rename users to user app, add club app
