+++ app/app/settings.py
@@ -47,7 +47,7 @@ INSTALLED_APPS = [
     'rest_framework.authtoken',
     'drf_spectacular',  # used for testing
     "core",
-    "users",
+    "user",
 ]
 
 MIDDLEWARE = [
