+++ app/setup.cfg
@@ -23,7 +23,7 @@ sort = Cover
 [flake8]
 max-line-length = 120
 exclude =
-  migrations
+  migrations/*,
   __pycache__,
   manage.py,
   settings.py
