+++ app/user/admin.py
@@ -4,7 +4,7 @@ Users admin config.
 
 from django.contrib import admin
 
-from users.models import User
+from user.models import User
 
 
 admin.site.register(User)
