+++ app/user/views.py
@@ -7,7 +7,7 @@ from rest_framework import generics, authentication, permissions
 from rest_framework.authtoken.views import ObtainAuthToken
 from rest_framework.settings import api_settings
 
-from users.serializers import (
+from user.serializers import (
     UserSerializer,
     AuthTokenSerializer,
 )
